+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Release NEXT:
  - Change usage of `OidcUserInfo` in interfaces to `OidcUserInfoExtended`, to also deserialize unknown properties
  - OID4VCI: `WalletService`: Replace parameters containing whole authentication parameters with single parameters holding `code` and `state`
  - Change several integer properties to durations, e.g. expirations (in seconds) for OIDC data classes
+ - In `SupportedCredentialFormat` replace `claims` with `isoClaims` and `sdJwtClaims` to be able to handle both formats defined in OID4VCI Draft 13
 
 Release 3.6.1:
  * Update to KMP-Crypto 2.6.0